refactor(parser): move FortranObj AST node
